+++ Core.cs
@@ -20,17 +20,17 @@ internal class Core
     static Entity _localCharacter = Entity.Null;
     static Entity _localUser = Entity.Null;
     public static Entity LocalCharacter =>
-        _localCharacter != Entity.Null
+        _localCharacter.Exists()
         ? _localCharacter
-        : (ConsoleShared.TryGetLocalCharacterInCurrentWorld(out _localCharacter, _client)
-        ? _localCharacter
-        : Entity.Null);
+        : (ConsoleShared.TryGetLocalCharacterInCurrentWorld(out _localCharacter, _client) && _localCharacter.Exists()
+            ? _localCharacter
+            : Entity.Null);
     public static Entity LocalUser =>
-        _localUser != Entity.Null
+        _localUser.Exists()
         ? _localUser
-        : (ConsoleShared.TryGetLocalUserInCurrentWorld(out _localUser, _client)
-        ? _localUser
-        : Entity.Null);
+        : (ConsoleShared.TryGetLocalUserInCurrentWorld(out _localUser, _client) && _localUser.Exists()
+            ? _localUser
+            : Entity.Null);
     public static EntityManager EntityManager => _client.EntityManager;
     public static ClientScriptMapper ClientScriptMapper { get; set; }
     public static ClientGameManager ClientGameManager { get; set; }